requirements.txt
torch
numpy
tqdm
pyyaml
pytorch-wavelets
